Move global dead-vector declaration to ConfigStandalone (out of mkFit.cc), move structs BeamSpot and DeadRegion out of Hit.h.
