+++ RecoTracker/MkFitCMS/standalone/Shell.h
@@ -3,6 +3,7 @@
 
 #include "RecoTracker/MkFitCore/interface/Hit.h"
 #include "RecoTracker/MkFitCore/interface/Track.h"
+#include "RecoTracker/MkFitCore/interface/DeadRegion.h"
 
 #include <map>
