+++ RecoTracker/MkFitCMS/standalone/mkFit.cc
@@ -41,11 +41,6 @@ using namespace mkfit;
 
 //==============================================================================
 
-namespace mkfit::internal {
-  // Filled in geometry plugin.
-  std::vector<DeadVec> deadvectors;
-}  // namespace mkfit::internal
-
 void initGeom() {
   std::cout << "Constructing geometry '" << Config::geomPlugin << "'\n";
